feat: centralize env and auth helpers
