+++ apps/crm-orchestrator/package.json
@@ -12,6 +12,7 @@
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "jsonwebtoken": "^9.0.2",
-    "mongodb": "^6.18.0"
+    "mongodb": "^6.18.0",
+    "@shared": "file:../../packages/shared"
   }
 }
