+++ apps/server/package.json
@@ -15,6 +15,7 @@
     "express-rate-limit": "^8.0.1",
     "jsonwebtoken": "^9.0.2",
     "socket.io": "^4.8.1",
-    "twilio": "^5.8.0"
+    "twilio": "^5.8.0",
+    "@shared": "file:../../packages/shared"
   }
 }
