+++ packages/shared/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "shared",
+  "version": "0.1.0",
+  "type": "module"
+}
